move Credential base class from cfme to cfme.base
