+++ cfme/tests/cloud_infra_common/test_vm_ownership.py
@@ -1,7 +1,8 @@
 import cfme.configure.access_control as ac
 import fauxfactory
 import pytest
-from cfme import Credential, login, test_requirements
+from cfme import login, test_requirements
+from cfme.base import Credential
 from cfme.common.vm import VM
 from utils import testgen
 from utils.blockers import BZ
