+++ cfme/tests/integration/test_cfme_auth.py
@@ -2,7 +2,8 @@
 import pytest
 
 from cfme.configure.access_control import Group, User
-from cfme import Credential, login
+from cfme import login
+from cfme.base import Credential
 from utils import browser
 from utils.conf import cfme_data
